backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pymupdf
pdfminer.six
pytesseract
Pillow
openai
langchain
langchain-openai
langsmith
faiss-cpu
numpy
pydantic
sqlalchemy
aiosqlite
python-json-logger
textstat
asyncio
httpx